@@ -1,12 +1,11 @@
-import { Request } from 'express'
 import constants from './constants.config'
-const moesif = require('moesif-nodejs')
+import moesif from 'moesif-nodejs'
 
 const moesifLoggingMiddleware = moesif({
     applicationId: constants.moesif.apiKey,
     logBody: false,
-    identifyUser: (req: Request) => req.session.userID || 'guest',
-    skip: (req: Request) => req.path === '/'
+    identifyUser: (req: object, _) => (req as { session: { userID: string | null | undefined } }).session.userID || 'guest',
+    skip: (req: object, _) => (req as { path: string | null | undefined }).path === '/'
 })
 
 export default moesifLoggingMiddleware
